Fixed bug on arm64 platforms for Android
Reviewed-on: https://wnproject.org/gerrit/71
Try: devpow112 <devon.f.powell@gmail.com>
Tested-by: WNBot <wn.bot.1@gmail.com>

+++ Overlays/Posix/Overlays/Android/Libraries/WNLogging/inc/Internal/WNConsoleLoggerAndroid.inl
@@ -10,7 +10,6 @@
 #include "WNLogging/inc/WNConsoleLogger.h"
 #include "WNUtilities/inc/WNLoggingData.h"
 
-
 namespace wn {
 namespace logging {
 
@@ -27,30 +26,29 @@ console_logger<T_Location>::~console_logger() {}
 
 template <console_location T_Location>
 void console_logger<T_Location>::flush_buffer(const char* _buffer,
-    size_t _buffer_size,
-    const color_element* _colors, size_t _num_colors) {
+    size_t _buffer_size, const color_element* _colors, size_t _num_colors) {
   if (_num_colors) {
     for (size_t i = 0; i < _num_colors; ++i) {
       if (static_cast<uint32_t>(_colors[i].m_level) > 0) {
         m_log_priority = static_cast<uint32_t>(_colors[i].m_level);
       } else {
         const char* end_color = (_num_colors == i + 1)
-                                   ? _buffer + _buffer_size
-                                   : _colors[i + 1].m_position;
+                                    ? _buffer + _buffer_size
+                                    : _colors[i + 1].m_position;
         size_t len = end_color - _colors[i].m_position;
-        __android_log_print(s_log_levels[static_cast<uint32_t>(m_log_priority)], wn::utilities::gAndroidLogTag, "%.*s",
-            static_cast<int32_t>(len), _colors[i].m_position);
+        __android_log_print(s_log_levels[static_cast<uint32_t>(m_log_priority)],
+            wn::utilities::gAndroidLogTag, "%.*s", static_cast<int32_t>(len),
+            _colors[i].m_position);
       }
     }
   } else {
-    __android_log_print(
-        s_log_levels[static_cast<uint32_t>(m_log_priority)], wn::utilities::gAndroidLogTag, "%.*s", (_buffer_size), _buffer);
+    __android_log_print(s_log_levels[static_cast<uint32_t>(m_log_priority)],
+        wn::utilities::gAndroidLogTag, "%.*s", static_cast<int>(_buffer_size),
+        _buffer);
   }
 }
 
-
-} // namespace logging
-} // namespace wn
-
+}  // namespace logging
+}  // namespace wn
 
 #endif  //__WN_CONSOLE_LOGGER_LINUX_INL__
